finished modularizing
